+++ framework/index.ts
@@ -39,8 +39,6 @@ export * from './PageTable/PageTableColumn';
 export * from './PageTable/PageToolbar';
 export * from './PageTable/useTableItems';
 export * from './PageTabs';
-export * from './RoutedTabs';
-export * from './PageBackTab';
 export * from './Settings';
 export * from './components/BulkSelector';
 export * from './components/Collapse';
