+++ frontend/awx/access/organizations/Organizations.tsx
@@ -39,13 +39,11 @@ import { AccessNav } from '../common/AccessNav';
 import { useSelectUsersAddOrganizations } from '../users/hooks/useSelectUsersAddOrganizations';
 import { useSelectUsersRemoveOrganizations } from '../users/hooks/useSelectUsersRemoveOrganizations';
 import { useDeleteOrganizations } from './hooks/useDeleteOrganizations';
-import { usePersistentFilters } from '../../../../framework/PersistentFilters';
 
 export function Organizations() {
   const { t } = useTranslation();
   const product: string = process.env.PRODUCT ?? t('AWX');
   const navigate = useNavigate();
-  usePersistentFilters('organizations');
 
   const toolbarFilters = useOrganizationsFilters();
