+++ frontend/awx/access/users/UserPage/UserPage.tsx
@@ -11,13 +11,11 @@ import {
   PageActions,
   PageHeader,
   PageLayout,
-  RoutedTabs,
-  RoutedTab,
-  PageBackTab,
 } from '../../../../../framework';
 import { LoadingPage } from '../../../../../framework/components/LoadingPage';
 import { RouteObj } from '../../../../Routes';
 import { useGetItem } from '../../../../common/crud/useGetItem';
+import { RoutedTabs, RoutedTab, PageBackTab } from '../../../../common/RoutedTabs';
 import { AwxError } from '../../../common/AwxError';
 import { User } from '../../../interfaces/User';
 import { useDeleteUsers } from '../hooks/useDeleteUsers';
